implement autoconfig for vitest reporter
