+++ .yarnrc.yml
@@ -16,6 +16,7 @@ packageExtensions:
   allure-js-commons@*:
     dependencies:
       allure-playwright: "workspace:*"
+      allure-vitest: "workspace:*"
 
 plugins:
   - ./plugin-ci-version.js
