+++ packages/allure-vitest/src/autoconfig.ts
@@ -1,0 +1,4 @@
+import { setGlobalTestRuntime } from "allure-js-commons/sdk/node";
+import { AllureVitestTestRuntime } from "./runtime.js";
+
+setGlobalTestRuntime(new AllureVitestTestRuntime());
