+++ packages/allure-vitest/src/runtime.ts
@@ -1,0 +1,188 @@
+/* eslint @typescript-eslint/require-await: off */
+import { type TaskContext, afterAll, afterEach, beforeAll, beforeEach } from "vitest";
+import {
+  ContentType,
+  Label,
+  LabelName,
+  Link,
+  LinkType,
+  MessagesHolder,
+  ParameterMode,
+  ParameterOptions,
+  RuntimeMessage,
+  Stage,
+  Status,
+  TestRuntime,
+} from "allure-js-commons/sdk/node";
+
+export class AllureVitestTestRuntime implements TestRuntime {
+  messagesHolder = new MessagesHolder();
+
+  async label(name: LabelName | string, value: string) {
+    this.sendMessage({
+      type: "metadata",
+      data: {
+        labels: [{ name, value }],
+      },
+    });
+  }
+
+  async labels(...labels: Label[]) {
+    this.sendMessage({
+      type: "metadata",
+      data: {
+        labels,
+      },
+    });
+  }
+
+  async link(url: string, type?: LinkType | string, name?: string) {
+    await this.sendMessage({
+      type: "metadata",
+      data: {
+        links: [{ type, url, name }],
+      },
+    });
+  }
+
+  async links(...links: Link[]) {
+    await this.sendMessage({
+      type: "metadata",
+      data: {
+        links,
+      },
+    });
+  }
+
+  async parameter(name: string, value: string, options?: ParameterOptions) {
+    await this.sendMessage({
+      type: "metadata",
+      data: {
+        parameters: [
+          {
+            name,
+            value,
+            ...options,
+          },
+        ],
+      },
+    });
+  }
+
+  async description(markdown: string) {
+    await this.sendMessage({
+      type: "metadata",
+      data: {
+        description: markdown,
+      },
+    });
+  }
+
+  async descriptionHtml(html: string) {
+    await this.sendMessage({
+      type: "metadata",
+      data: {
+        descriptionHtml: html,
+      },
+    });
+  }
+
+  async displayName(name: string) {
+    await this.sendMessage({
+      type: "metadata",
+      data: {
+        displayName: name,
+      },
+    });
+  }
+
+  async historyId(value: string) {
+    await this.sendMessage({
+      type: "metadata",
+      data: {
+        historyId: value,
+      },
+    });
+  }
+
+  async testCaseId(value: string) {
+    await this.sendMessage({
+      type: "metadata",
+      data: {
+        testCaseId: value,
+      },
+    });
+  }
+
+  async attachment(name: string, content: Buffer | string, type: string | ContentType) {
+    await this.sendMessage({
+      type: "raw_attachment",
+      data: {
+        name,
+        content: Buffer.from(content).toString("base64"),
+        contentType: type,
+        encoding: "base64",
+      },
+    });
+  }
+
+  async step(name: string, body: () => void | PromiseLike<void>) {
+    await this.sendMessage({
+      type: "step_start",
+      data: {
+        name,
+        start: Date.now(),
+      },
+    });
+
+    try {
+      await body();
+
+      await this.sendMessage({
+        type: "step_stop",
+        data: {
+          status: Status.PASSED,
+          stage: Stage.FINISHED,
+          stop: Date.now(),
+        },
+      });
+    } catch (err) {
+      const { message, stack } = err as Error;
+
+      await this.sendMessage({
+        type: "step_stop",
+        data: {
+          status: Status.FAILED,
+          stage: Stage.FINISHED,
+          stop: Date.now(),
+          statusDetails: {
+            message,
+            trace: stack,
+          },
+        },
+      });
+
+      throw err;
+    }
+  }
+
+  async stepDisplayName(name: string) {
+    await this.sendMessage({
+      type: "step_metadata",
+      data: { name },
+    });
+  }
+
+  async stepParameter(name: string, value: string, mode?: ParameterMode) {
+    await this.sendMessage({
+      type: "step_metadata",
+      data: {
+        parameters: [{ name, value, mode }],
+      },
+    });
+  }
+
+  async sendMessage(message: RuntimeMessage) {
+    this.messagesHolder.push(message);
+  }
+}
